Smart Pointers and King cannot walk into check:
I have finished converting the engine to use smart pointers except for one bug with the Window pointer in the Application class.

The second todo is also almost done because the king cannot walk into check. All that is left to do is recalculate the piece moves if the king is in check.

TODO: recalculate piece moves after calculating king moves if king is in check, so pieces can block king check